Rename solve_brute → backward_induction (module + tests)
The solve module's job is backward induction; name it for that. Renames
src/_lcm/solution/solve_brute.py → backward_induction.py and
tests/solution/test_solve_brute.py → test_backward_induction.py, updates
every import (model.py, simulation/compile.py, test_beartype_claw.py),
the two test-function names, and the doc/docstring references.
