+++ src/_lcm/simulation/compile.py
@@ -25,7 +25,7 @@ from dags.tree import qname_from_tree_path
 from _lcm.engine import Regime
 from _lcm.simulation.initial_conditions import subject_array_sharding
 from _lcm.simulation.random import generate_simulation_keys
-from _lcm.solution.solve_brute import (
+from _lcm.solution.backward_induction import (
     _build_zero_V_arr,
     _func_dedup_key,
     _get_regime_V_shapes_and_shardings,
