+++ src/_lcm/solution/backward_induction.py
